ILoveDocuments/files/rag-system/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic-settings==2.2.1
chromadb==0.5.0
openai==1.30.0
langchain-text-splitters==0.2.0
pypdf==4.2.0
python-docx==1.1.0
tiktoken==0.7.0
httpx==0.27.0
